refactor(cloud-storage): introduce useCloudStorageBrowser factory + CloudFileBrowserShell, migrate Nextcloud
Unit 2 commit 1 of 4 from concepts/2026-05-11 Cloud Storage Browser
Shell Extraction.md. Adds two new files:

- client/src/features/upload/hooks/useCloudStorageBrowser.js — state
  management factory parameterised by an adapter that captures the
  navigation model (id-based Graph/Drive vs path-based WebDAV),
  query/breadcrumb shape, and provider-specific predicates (e.g.
  Google's virtual `isGoogleDoc` items).
- client/src/features/upload/components/CloudFileBrowserShell.jsx —
  shared render shell covering the not-connected / sources / drives /
  files four-view flow, sort + filter + search, selection + download
  progress, attach affordance.

Migrates Nextcloud to use them: the hook shrinks from 318 lines to 32
and the component from 568 lines to 25. UX is unchanged, including the
single-drive auto-skip (Nextcloud has exactly one drive per user) and
the previously-cleaned-up "Sorted by" debug indicator (still absent).

Google Drive and Office 365 still use their own hooks/components in
this commit. They are migrated one-per-commit in the next two
commits so each commit ships green.
